projeto rdiante plataforma/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
firebase-admin==6.2.0
pyrebase4==4.7.0 